Rename example resource to `app`
