+++ service/controller/resource/app/create.go
@@ -1,4 +1,4 @@
-package test
+package app
 
 import (
 	"context"
